move unit tests into test/unit
This makes room for the fuzz testing framework.
