+++ test/unit/Entry.js
