+++ test/unit/Log.js
