+++ test/unit/LogEntryApplier.js
