+++ test/unit/MessageBuffer.js
