+++ test/unit/Peer.js
