+++ test/unit/State.js
@@ -2,7 +2,7 @@ import { beforeEach, context, describe, it } from '!mocha'
 import assert from 'power-assert'
 import { stub } from 'sinon'
 
-import { getReason } from './support/utils'
+import { getReason } from '../support/utils'
 
 import State from '🏠/lib/State'
