+++ test/unit/expose-events.js
